rename files archivied-places for jsdoc
