+++ server/src/models/archived-place/index.js
@@ -1,1 +1,1 @@
-export * from './archived-place.queries'
+export * from './archived-place-queries'
